Add displaying a list of products.

+++ SportShop.Context.Data/Migrations/20221112205930_Initial.cs
@@ -36,7 +36,7 @@ namespace SportShop.Context.Data.Migrations
                 name: "Product",
                 columns: table => new
                 {
-                    ProductId = table.Column<long>(type: "bigint", nullable: false)
+                    ProductId = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Description = table.Column<string>(type: "nvarchar(max)", nullable: false),
